+++ tests/MWC.Core.Tests/ApplePhase3Tests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
+using MWC.App.Services;    // UpdateCheckResult
 using MWC.Core.Services;
 using Xunit;